Fix bugs of wrong usage of mockMobile function in test
In testing, the helper function mockMobile is to mockup the browser,
to make the browser acts like on mobile device, such as iOS and Android.
The common usage of mockMobile is:

1) Call it and store the return value to a variable // const mock = mockMobile();
2) execute the testing code, such as assertions
3) Call restore method to stop mocking-up // mock.restore();

It works fine if the assertions passed, but if any assertion failed, step 3
will NOT executed. Due to the limitation of mockMobile, the function does not
provide mocked-up sandbox environment to each test. Instead, it mocks-up the
whole browser. If restore method is not executed, the environments of other
test cases may be effected.

The commit resolves the problem by wrapping step 2 in a try block, and executing
step 3 in the finally block. This ensures step 3 will be executed no matter the
assertions failed or not.
The following code illustrates the modified usage:

const mock = mockMobile();
try {
  // Testing code here
} finally {
  mock.restore();
}