Transform constants/events from static values to function
The constant "events" contains static values that can be different on
desktop and mobile. (For exmaple, "mousedown" on desktop and "touchstart"
on mobile.) The original implementation may have some problems in testing,
because the values are determined when "events" moudle is imported. In
testing, If we mockup mobile device after importing "events" (or modules
that import "events"), the values remain the same. I resolve the problem
by making "events" as a function that determing return value when calling
it.