Replace PhantomJS with SlimerJS
Panophoto uses THREE.js as a library to manipulate WebGL. The
Problem of using PhatomJS as the testing browser is it doesn't
support WebGL API. So we choose an alternative browser: SlimerJS,
which supports WebGL well. Although it is not a true headless
browser, but it is still convenience for developing.