Minimize the code of bcryptjs and Bluebird
- Minimize bcryptjs:
When we use bcryptjs, the built file size increases dramatically
(~100KB in gzip) bacause Webpack detects it may uses crypto
library and load the whole browserified version of crypto. To
prevent from loading the whole crypto library, we maintained our
bcryptjs which does not require it. The size of bcryptjs in built
file is only about 10KB in gzip.

- Minimize Bluebird
We use promisify function in Bluebird and the size of Bluebird in
built file size is about 20KB. To reduce the size, we use our own
promisify function instead of Bluebird.